streamlit-pdf-merger/requirements.txt
streamlit
PyPDF2
Pillow
PyMuPDF